backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
gitpython>=3.1.40
openai>=1.6.0
pydantic>=2.5.0
